requirements.txt
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
Faker==37.3.0
fonttools==4.58.2
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
tzdata==2025.2
